requirements.txt
fastapi
uvicorn
boto3
python-dotenv
aiohttp
numpy
faiss-cpu
pypdf
playwright
fredapi
dbnomics
wbdata
yfinance
pandas
openpyxl
xlsxwriter
tqdm
colorama
scipy
statsmodels
